Refactor naming (#105)
* Refactor naming
* Delete old ds model trainer
