+++ knodle/trainer/__init__.py
@@ -1,1 +1,1 @@
-from knodle.trainer.config.trainer_config import TrainerConfig
+from knodle.trainer.config import TrainerConfig
