+++ knodle/trainer/baseline/bert.py
@@ -7,7 +7,7 @@ import torch
 from torch import Tensor
 from torch.utils.data import TensorDataset
 
-from knodle.trainer.ds_model_trainer.ds_model_trainer import DsModelTrainer
+from knodle.trainer.ds_model_trainer.ds_model_trainer import Trainer
 from knodle.trainer.utils import log_section
 from knodle.trainer.utils.denoise import get_majority_vote_probs
 from knodle.trainer.utils.utils import accuracy_of_probs
@@ -15,7 +15,7 @@ from knodle.trainer.utils.utils import accuracy_of_probs
 logger = logging.getLogger(__name__)
 
 
-class MajorityBertTrainer(DsModelTrainer):
+class MajorityBertTrainer(Trainer):
     def train(self):
         """
         This function gets final labels with a majority vote approach and trains the provided model.
