+++ knodle/trainer/crossweigh_weighing/crossweigh_denoising_config.py
@@ -8,28 +8,29 @@ from torch.optim import optimizer
 
 
 class CrossWeighDenoisingConfig:
-    def __init__(self,
-                 model: nn.Module,
-                 crossweigh_partitions: int = 3,
-                 crossweigh_folds: int = 5,
-                 crossweigh_epochs: int = 2,
-                 weight_reducing_rate: int = 0.5,
-                 samples_start_weights: int = 3.0,
-                 no_relation_weights: int = 0.5,
-                 size_factor: int = 200,
-                 batch_size: int = 32,
-                 lr: float = 0.01,
-                 output_classes: int = 2,
-                 class_weights: Tensor = None,
-                 optimizer_: optimizer = None,
-                 criterion: Callable[[Tensor, Tensor], float] = cross_entropy_with_probs,
-                 path_to_weights: str = "data/sample_weights",
-                 seed: int = "12345",
-                 enable_cuda: bool = False,
-                 use_grad_clipping: bool = True,
-                 grad_clipping: int = 5,
-                 negative_samples: bool = False     # True if there is a special label for no_relation samples
-                 ):
+    def __init__(
+            self,
+            model: nn.Module,
+            crossweigh_partitions: int = 3,
+            crossweigh_folds: int = 5,
+            crossweigh_epochs: int = 2,
+            weight_reducing_rate: int = 0.5,
+            samples_start_weights: int = 3.0,
+            no_relation_weights: int = 0.5,
+            size_factor: int = 200,
+            batch_size: int = 32,
+            lr: float = 0.01,
+            output_classes: int = 2,
+            class_weights: Tensor = None,
+            optimizer_: optimizer = None,
+            criterion: Callable[[Tensor, Tensor], float] = cross_entropy_with_probs,
+            path_to_weights: str = "data/sample_weights",
+            seed: int = "12345",
+            enable_cuda: bool = False,
+            use_grad_clipping: bool = True,
+            grad_clipping: int = 5,
+            negative_samples: bool = False  # True if there is a special label for no_relation samples
+    ):
 
         self.cw_partitions = crossweigh_partitions
         self.cw_folds = crossweigh_folds
@@ -62,5 +63,3 @@ class CrossWeighDenoisingConfig:
             self.optimizer = torch.optim.Adam(model.parameters(), lr=self.lr)
         else:
             self.optimizer = optimizer_
-
-
