+++ knodle/trainer/knn_tfidf_similarities/knn_tfidf_similarity.py
@@ -9,7 +9,7 @@ from torch.utils.data import TensorDataset, DataLoader
 from tqdm import tqdm
 
 from knodle.trainer import TrainerConfig
-from knodle.trainer.ds_model_trainer.ds_model_trainer import DsModelTrainer
+from knodle.trainer.trainer import Trainer
 from knodle.trainer.utils import log_section
 from knodle.trainer.utils.denoise import get_majority_vote_probs
 from knodle.trainer.utils.utils import accuracy_of_probs, extract_tensor_from_dataset
@@ -17,7 +17,7 @@ from knodle.trainer.utils.utils import accuracy_of_probs, extract_tensor_from_da
 logger = logging.getLogger(__name__)
 
 
-class KnnTfidfSimilarity(DsModelTrainer):
+class KnnTfidfSimilarity(Trainer):
     def __init__(
         self,
         model: Module,
